# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Scheduling Service API
  version: '0.1.0'
  description: >-
    API responsável por criar, cancelar e gerenciar agendamentos para as unidades da barbearia.
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
paths:
  /agendamentos:
    post:
      summary: Criar agendamento
      tags:
        - Agendamentos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
            examples:
              default:
                value:
                  clientId: "ef8730bd-8dbe-45b6-b5b4-2cb4c2ff01d8"
                  unitId: "a3d932ae-f23c-4ce5-9ed3-2f2eca4a0df8"
                  serviceId: "5b6ad154-42d8-4a73-8c23-5ce77e9d0b74"
                  barberId: "e5a7b492-991a-4e54-892d-8c4191dcf689"
                  start: "2025-12-03T15:00:00Z"
                  reservationToken: "resv_tok_abcdef123456"
                  origin: "cliente"
                  notes: "Primeira visita"
      responses:
        '201':
          description: Agendamento criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResource'
        '409':
          description: Reserva já utilizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Listar agendamentos
      tags:
        - Agendamentos
      parameters:
        - in: query
          name: clienteId
          schema:
            type: string
            format: uuid
        - in: query
          name: barberId
          schema:
            type: string
            format: uuid
        - in: query
          name: unitId
          schema:
            type: string
            format: uuid
        - in: query
          name: date
          schema:
            type: string
            format: date
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista paginada de agendamentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAppointmentsResponse'
  /agendamentos/{id}:
    get:
      summary: Obter agendamento por ID
      tags:
        - Agendamentos
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Agendamento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResource'
        '404':
          description: Agendamento não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /agendamentos/{id}/cancel:
    put:
      summary: Cancelar agendamento
      tags:
        - Agendamentos
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelAppointmentRequest'
            examples:
              default:
                value:
                  reason: "Cliente solicitou cancelamento"
      responses:
        '200':
          description: Agendamento cancelado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResource'
  /agendamentos/{id}/falta:
    put:
      summary: Marcar falta
      tags:
        - Agendamentos
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkNoShowRequest'
            examples:
              default:
                value:
                  timestamp: "2025-12-03T15:45:00Z"
      responses:
        '200':
          description: Agendamento marcado como falta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResource'
  /auth/login:
    post:
      summary: Login de barbeiro/admin
      tags:
        - Autenticação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Tokens gerados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
  /auth/refresh:
    post:
      summary: Renovar tokens
      tags:
        - Autenticação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRefreshRequest'
      responses:
        '200':
          description: Tokens renovados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
  /slots/lock:
    post:
      summary: Exemplo de lock de slot (serviço Availability)
      description: Payload fornecido como referência para integração com Availability Service.
      tags:
        - Referências
      requestBody:
        required: true
        content:
          application/json:
            examples:
              default:
                value:
                  unitId: "7a6b5c4d-3e2f-1a0b-9c8d-7e6f5a4b3c2d"
                  serviceId: "3f2e1d0c-9b8a-7f6e-5d4c-3b2a1f0e9d8c"
                  barberId: null
                  start: "2025-12-03T15:00:00Z"
                  end: "2025-12-03T15:30:00Z"
                  clientRequestId: "req-987654321"
      responses:
        '200':
          description: Token de reserva criado
          content:
            application/json:
              examples:
                default:
                  value:
                    reservationToken: "resv_tok_abcdef123456"
                    expiresInSeconds: 120
components:
  schemas:
    AppointmentResource:
      type: object
      required:
        - appointmentId
        - clientId
        - unitId
        - serviceId
        - start
        - end
        - status
        - origin
        - createdAt
        - updatedAt
      properties:
        appointmentId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        unitId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        barberId:
          type: string
          format: uuid
          nullable: true
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        status:
          type: string
          enum: [agendado, cancelado, falta]
        origin:
          type: string
          enum: [cliente, barbeiro, admin]
        reservationToken:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAppointmentRequest:
      type: object
      required:
        - clientId
        - unitId
        - serviceId
        - start
        - reservationToken
        - origin
      properties:
        clientId:
          type: string
          format: uuid
        unitId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        barberId:
          type: string
          format: uuid
          nullable: true
        start:
          type: string
          format: date-time
        reservationToken:
          type: string
        origin:
          type: string
          enum: [cliente, barbeiro, admin]
        notes:
          type: string
          nullable: true
    ListAppointmentsResponse:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AppointmentResource'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    CancelAppointmentRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          maxLength: 280
    MarkNoShowRequest:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
    AuthLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    AuthRefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    AuthTokensResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - user
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          type: object
          required:
            - id
            - email
            - role
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            role:
              type: string
              enum: [barbeiro, admin]
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT